reorganize codebase, add fonts, replace firestore images
