+++ src/App.css
@@ -1,6 +1,15 @@
 @import url("https://fonts.googleapis.com/css2?family=Didact+Gothic&display=swap");
 @import "tailwindcss";
 
+@font-face {
+    font-family: 'OrganGrinder';
+    src: url('../src/assets/fonts/OrganGrinder.woff2') format('woff2'),
+        url('../src/assets/fonts/OrganGrinder.woff') format('woff');
+    font-weight: normal;
+    font-style: normal;
+    font-display: swap;
+}
+
 #root {
   max-width: 1280px;
   margin: 0 auto;
