+++ src/App.jsx
@@ -4,7 +4,7 @@ import Landing from "./pages/Landing/Landing.jsx";
 import Login from "./pages/Login.jsx";
 import Signup from "./pages/Signup";
 import Results from "./components/Results.jsx";
-import Assessment from "./pages/Assessment.jsx";
+import Assessment from "./pages/Assessment/Assessment.jsx";
 import "./css/colors.css";
 
 function App() {
