--- src/css/QuizBar.module.css
@@ -1,23 +1,0 @@
-.completed {
-    width: 10px;
-    height: 10px;
-    margin: 5px;
-    border-radius: 50%;
-    background-color: #C09726;
-    border-radius: 1px solid black;
-}
-
-.remaining {
-    width: 10px;
-    height: 10px;
-    margin: 5px;
-    border-radius: 50%;
-    background-color: #650909;
-    border-radius: 1px solid black;
-}
-
-.row {
-    display: flex;
-    justify-content: center;
-    padding: 10px;
-}
