--- src/css/QuizTitle.module.css
@@ -1,7 +1,0 @@
-/* OLD FILE */
-
-.title {
-    font-size: 2em;
-    font-weight: 700;
-    margin-bottom: 10px;
-}
