+++ src/pages/Assessment/BreakButton/BreakButton.module.css
@@ -8,7 +8,7 @@ dialog {
   padding: 1rem;
   border-radius: 15px;
   width: 50vw;
-  height: 30vh;
+  height: 35vh;
 }
 
 dialog::backdrop {
@@ -48,7 +48,8 @@ dialog::backdrop {
     background-color: var(--main-mustard);
 }
 .dialogTitle {
-    font-size: 2em;
+    font-family: "OrganGrinder";
+    font-size: 3em;
 }
 
 .dialogText {
@@ -63,7 +64,7 @@ dialog::backdrop {
 
 .dialogIcon {
     z-index: 1;
-    width: 5em;
-    height: 5em;
+    width: 100%;
+    height: 100%;
     
 }
