+++ src/pages/Assessment/NavigationButton/NavigationButton.module.css
@@ -40,9 +40,8 @@
     background-color: #FFF9F2;
 }
 
-.skipIcon {
-    width: 70%;
-    height: 70%;
-    padding-left: 5px;
+.icon {
+  width: 100%;
+  height: 100%;
 }
