+++ src/pages/Assessment/QuestionTypes/MultiselectImage.jsx
@@ -1,16 +1,16 @@
-import {doc, getDoc} from "firebase/firestore"; 
-import {useEffect, useState} from "react";
-import {db} from "../../firebase/config";
-import { getDownloadURL, getStorage, ref } from "firebase/storage";
-import styles from "../../css/Assessment/SelectImage.module.css";
-import ToggleSelection from "./ToggleSelection";
+import { doc, getDoc } from "firebase/firestore";
+import { useEffect, useState } from "react";
+import { db } from "../../../firebase/config";
+// import { getDownloadURL, getStorage, ref } from "firebase/storage";
+import check from "../assets/check-icon.svg";
+import styles from "./SelectImage.module.css";
+import ToggleSelection from "../components/ToggleSelection";
 
-const MultiselectImage = ({ formValues, setFormValues, sectionId, id}) => {
+const MultiselectImage = ({ formValues, setFormValues, sectionId, id }) => {
     const [options, setOptions] = useState([]);
     const [description, setDescription] = useState("");
     const [select, setSelection] = useState(0);
-    const [checkIcon, setCheckURL] = useState(null);
-    const [isLoading, setIsLoading] = useState(true);
+    // const [checkIcon, setCheckURL] = useState(null);
     const selected = formValues[id] || [];
 
     // Modular Firebase Firestore query
@@ -22,7 +22,7 @@ const MultiselectImage = ({ formValues, setFormValues, sectionId, id}) => {
             // Fetches document data
             const docSnap = await getDoc(docRef);
 
-            if(docSnap.exists()) {
+            if (docSnap.exists()) {
                 setOptions(docSnap.data().options || []);
                 setDescription(docSnap.data().description || "");
                 setSelection(docSnap.data().select || 2);
@@ -32,7 +32,7 @@ const MultiselectImage = ({ formValues, setFormValues, sectionId, id}) => {
         };
         fetchOptions();
     }, [id]);
-
+/*
     useEffect(() => {
         const storage = getStorage();
         const checkRef = ref(storage, "icons/check.png");
@@ -40,7 +40,7 @@ const MultiselectImage = ({ formValues, setFormValues, sectionId, id}) => {
         getDownloadURL(checkRef)
             .then(url => setCheckURL(url))
             .catch(error => console.error("Failed to fetch icon, ", error))
-    }, []);
+    }, []);*/
 
     return (
         <>
@@ -48,19 +48,19 @@ const MultiselectImage = ({ formValues, setFormValues, sectionId, id}) => {
 
             <div className={styles.grid}>
                 {options.map((opt, idx) => (
-                    <div key={idx} className={styles.imageBorder} 
-                                    onClick={() => ToggleSelection(select, opt.label, setFormValues, id)}>
-                        
+                    <div key={idx} className={styles.imageBorder}
+                        onClick={() => ToggleSelection(select, opt.label, setFormValues, id)}>
+
                         <div className={styles.imageOption}>
                             <img src={opt.image} alt={opt.label} className={styles.image} />
                         </div>
 
                         <div className={styles.imageFooter}>
-                            <div className={styles.check}> 
+                            <div className={styles.check}>
                                 <div className={`${selected.includes(opt.label) ? styles.imageFilled : styles.imageSelect}`}></div>
-                                {selected.includes(opt.label) ? (<img src={checkIcon} className={styles.checkMark}/>) : ""}
+                                {selected.includes(opt.label) ? (<img src={check} className={styles.checkMark} />) : ""}
                             </div>
-                            
+
                             <div className={styles.imageCaption}>{opt.label}</div>
                             <div></div>
                         </div>
