+++ src/pages/Assessment/QuestionTypes/MultiselectText.jsx
@@ -1,11 +1,11 @@
-import {doc, getDoc} from "firebase/firestore";
-import {useEffect, useState} from "react";
-import {db} from "../../firebase/config";
-import styles from "../../css/Assessment/SelectText.module.css";
-import ToggleSelection from "./ToggleSelection";
-import OtherOption from "./OtherOption";
+import { doc, getDoc } from "firebase/firestore";
+import { useEffect, useState } from "react";
+import { db } from "../../../firebase/config";
+import styles from "./SelectText.module.css";
+import ToggleSelection from "../components/ToggleSelection";
+import OtherOption from "../components/OtherOption";
 
-const MultiselectText = ({formValues, setFormValues, sectionId, id}) => {
+const MultiselectText = ({ formValues, setFormValues, sectionId, id }) => {
     const [options, setOptions] = useState([]);
     const [select, setSelection] = useState(0);
     const [otherText, setOtherText] = useState("");
@@ -20,7 +20,7 @@ const MultiselectText = ({formValues, setFormValues, sectionId, id}) => {
             // Fetches document data
             const docSnap = await getDoc(docRef);
 
-            if(docSnap.exists()) {
+            if (docSnap.exists()) {
                 setOptions(docSnap.data().options || []);
                 setSelection(docSnap.data().select || 2);
                 setOtherText(docSnap.data().otherText || "Other");
@@ -37,7 +37,7 @@ const MultiselectText = ({formValues, setFormValues, sectionId, id}) => {
                 <div className={styles.multipleChoice}>
                     {options.map((opt, idx) => (
                         opt === "Other" ? (
-                            <OtherOption 
+                            <OtherOption
                                 key={idx}
                                 formValues={formValues}
                                 setFormValues={setFormValues}
@@ -49,7 +49,7 @@ const MultiselectText = ({formValues, setFormValues, sectionId, id}) => {
                             <div className={styles.textOption}
                                 onClick={() => ToggleSelection(select, opt, setFormValues, id)}>
                                 <div className={`${styles.optionIcon} ${selected.includes(opt) ? styles.filled : ""}`}></div>
-                                
+
                                 <div
                                     key={idx}
                                     className={styles.textOptionStyle}
