+++ src/pages/Assessment/QuestionTypes/SectionDivider.module.css
@@ -2,7 +2,8 @@
     padding: 2rem;
 }
 .section {
-    font-size: 2rem;
+    font-family: "OrganGrinder";
+    font-size: 4rem;
     color: white;
 }
