+++ src/pages/Assessment/QuestionTypes/SelectImage.module.css
